refactor downstream api to use timers, cleanup
Use new timers in the downstream API to have a better resolution for
events.

This commit also deletes phase_context from the library, as this way
of computing phase information was very rigid and not applicable to a
lot of use cases. A better approach would be to refactor our
sensor/event handling to be able to compute arbitrary formulas on top of
generic events.