working? event timeseries database
It maintain a hash of sensor uuids to a list of
scoped events. Scoped events associate a scope with a timeseries.
Inspired by rrdtool, the timeserie is actually stored in two
ringbuffers, one for previous periods of control (after a tick) one for
the current time window. The current time window events are aggregated
when the ringbuffer is full (which kind of defeat the point of a
ringbuffer, I agree).

Management of ticks and history of time windows will be done later.