[feature] add clientid to actuator
Have the nrm daemon save the client that added an actuator, to be able
to send it a message later when actuation is needed.

Of course for the cli it makes no sense, since the client is going to be
done by the time the daemon actually wants to communicate back.