[feature] add callback for actuate
As annoying as it is, there's no easy way to deal with async actuate
than with another callback on the client side.

What might be valuable though, is to start thinking about a way to
configure roles so that each role-level event can be configured to
trigger either a callback or an action on the pipes/sockets.